Migrate cos tests
