+++ src/unittests/f32_interval.spec.ts
@@ -20,7 +20,6 @@ import {
   clampMedianInterval,
   clampMinMaxInterval,
   correctlyRoundedInterval,
-  cosInterval,
   coshInterval,
   crossInterval,
   degreesInterval,
@@ -1923,42 +1922,6 @@ interface ScalarToIntervalCase {
   input: number;
   expected: number | IntervalBounds;
 }
-
-g.test('cosInterval')
-  .paramsSubcasesOnly<ScalarToIntervalCase>(
-    // prettier-ignore
-    [
-      // This test does not include some common cases. i.e. f(x = π/2) = 0,
-      // because the difference between true x and x as a f32 is sufficiently
-      // large, such that the high slope of f @ x causes the results to be
-      // substantially different, so instead of getting 0 you get a value on the
-      // order of 10^-8 away from 0, thus difficult to express in a
-      // human-readable manner.
-      { input: kValue.f32.infinity.negative, expected: kAnyBounds },
-      { input: kValue.f32.negative.min, expected: kAnyBounds },
-      { input: kValue.f32.negative.pi.whole, expected: [-1, plusOneULP(-1)] },
-      { input: kValue.f32.negative.pi.third, expected: [minusOneULP(1/2), 1/2] },
-      { input: 0, expected: [1, 1] },
-      { input: kValue.f32.positive.pi.third, expected: [minusOneULP(1/2), 1/2] },
-      { input: kValue.f32.positive.pi.whole, expected: [-1, plusOneULP(-1)] },
-      { input: kValue.f32.positive.max, expected: kAnyBounds },
-      { input: kValue.f32.infinity.positive, expected: kAnyBounds },
-    ]
-  )
-  .fn(t => {
-    const error = (_: number): number => {
-      return 2 ** -11;
-    };
-
-    t.params.expected = applyError(t.params.expected, error);
-    const expected = toF32Interval(t.params.expected);
-
-    const got = cosInterval(t.params.input);
-    t.expect(
-      objectEquals(expected, got),
-      `cosInterval(${t.params.input}) returned ${got}. Expected ${expected}`
-    );
-  });
 
 g.test('coshInterval')
   .paramsSubcasesOnly<ScalarToIntervalCase>(
