+++ src/unittests/floating_point.spec.ts
@@ -363,3 +363,39 @@ g.test('ceilInterval_f32')
       `f32.ceilInterval(${t.params.input}) returned ${got}. Expected ${expected}`
     );
   });
+
+g.test('cosInterval_f32')
+  .paramsSubcasesOnly<ScalarToIntervalCase>(
+    // prettier-ignore
+    [
+      // This test does not include some common cases. i.e. f(x = π/2) = 0,
+      // because the difference between true x and x as a f32 is sufficiently
+      // large, such that the high slope of f @ x causes the results to be
+      // substantially different, so instead of getting 0 you get a value on the
+      // order of 10^-8 away from 0, thus difficult to express in a
+      // human-readable manner.
+      { input: kValue.f32.infinity.negative, expected: kAnyBounds },
+      { input: kValue.f32.negative.min, expected: kAnyBounds },
+      { input: kValue.f32.negative.pi.whole, expected: [-1, plusOneULPF32(-1)] },
+      { input: kValue.f32.negative.pi.third, expected: [minusOneULPF32(1/2), 1/2] },
+      { input: 0, expected: [1, 1] },
+      { input: kValue.f32.positive.pi.third, expected: [minusOneULPF32(1/2), 1/2] },
+      { input: kValue.f32.positive.pi.whole, expected: [-1, plusOneULPF32(-1)] },
+      { input: kValue.f32.positive.max, expected: kAnyBounds },
+      { input: kValue.f32.infinity.positive, expected: kAnyBounds },
+    ]
+  )
+  .fn(t => {
+    const error = (_: number): number => {
+      return 2 ** -11;
+    };
+
+    t.params.expected = applyError(t.params.expected, error);
+    const expected = FP.f32.toInterval(t.params.expected);
+
+    const got = FP.f32.cosInterval(t.params.input);
+    t.expect(
+      objectEquals(expected, got),
+      `f32.cosInterval(${t.params.input}) returned ${got}. Expected ${expected}`
+    );
+  });
